Add missing functions to ETLService
